refactor: split optional deps into extras
flask, rpyc, pyyaml, requests are no longer required by `import registry`.
They become extras, loaded only when the relevant submodule is imported.

- pyproject.toml: deps = pydantic + typing-extensions + tomli;
  new extras yaml/http/rpc/server; python floor 3.8 -> 3.10
- __init__.py: drop _REQUIRED_DEPS gate and engines re-export
- storage.py: holds ThreadSafeLocalStorage only
- remote_storage.py (new): RemoteStorageProxy + helpers, imports requests
- typ_registry.py, fnc_registry.py: drop unused logic_namespace kwarg
- __main__.py: import RemoteStorageProxy from new location

All imports stay at module top. No lazy imports.

BREAKING:
- `from registry import ConfigFileEngine, SocketEngine` -> use `from registry.engines import ...`
- `TypeRegistry[T](..., logic_namespace="ns")` removed; assign _repository directly
- Drop Python 3.8/3.9
